+++ R/last_fit.R
@@ -4,16 +4,16 @@
 #' the final fit on the entire training set is needed and is then evaluated on
 #' the test set.
 #'
-#' @param object A workflow, formula, or recipe. No tuning parameters are allowed.
+#' @param object A `parsnip` model specification or a [workflows::workflow()].
+#'   No tuning parameters are allowed.
 #'
-#' @param model A `parsnip` model specification. No tuning parameters are allowed.
+#' @param preprocessor A traditional model formula or a recipe created using
+#'   [recipes::recipe()].
 #'
 #' @param split An `rsplit` object created from [rsample::initial_split()].
 #'
 #' @param metrics A [yardstick::metric_set()], or `NULL` to compute a standard
 #'   set of metrics.
-#'
-#' @param formula A formula specifying the terms of the model.
 #'
 #' @param ... Currently unused.
 #'
@@ -40,7 +40,7 @@
 #' lin_mod <- linear_reg() %>%
 #'   set_engine("lm")
 #'
-#' spline_res <- last_fit(spline_rec, lin_mod, split = tr_te_split)
+#' spline_res <- last_fit(lin_mod, spline_rec, split = tr_te_split)
 #' spline_res
 #'
 #' # test set results
@@ -65,41 +65,55 @@ last_fit <- function(object, ...) {
 last_fit.default <- function(object, ...) {
   empty_ellipses(...)
   msg <- paste0(
-    "The first argument to [last_fit()] should be either a ",
-    "formula, recipe, or workflow."
+    "The first argument to [last_fit()] should be either ",
+    "a model or workflow."
   )
   rlang::abort(msg)
 }
 
-#' @rdname last_fit
 #' @export
 last_fit.recipe <- function(object, model, split, ..., metrics = NULL) {
+  lifecycle::deprecate_soft("0.1.0",
+                            what = "last_fit.recipe()",
+                            details = deprecate_msg(match.call(), "last_fit"))
   empty_ellipses(...)
-  if (is_missing(model) || !inherits(model, "model_spec")) {
-    rlang::abort("`model` should be a parsnip model specification object.")
-  }
 
-  workflow <- workflow()
-  workflow <- add_recipe(workflow, object)
-  workflow <- add_model(workflow, model)
+  last_fit(model, preprocessor = object, split = split, metrics = metrics)
 
-  last_fit_workflow(workflow, split, metrics)
 }
 
-#' @rdname last_fit
 #' @export
 last_fit.formula <- function(formula, model, split, ..., metrics = NULL) {
+  lifecycle::deprecate_soft("0.1.0",
+                            what = "last_fit.formula()",
+                            details = deprecate_msg(match.call(), "last_fit"))
   empty_ellipses(...)
-  if (is_missing(model) || !inherits(model, "model_spec")) {
-    rlang::abort("`model` should be a parsnip model specification object.")
+
+  last_fit(model, preprocessor = formula, split = split, metrics = metrics)
+}
+
+#' @export
+#' @rdname last_fit
+last_fit.model_spec <- function(object, preprocessor, split, ..., metrics = NULL) {
+
+  if (rlang::is_missing(preprocessor) || !is_preprocessor(preprocessor)) {
+    rlang::abort(paste("To tune a model spec, you must preprocess",
+                       "with a formula or recipe"))
   }
 
-  workflow <- workflow()
-  workflow <- add_formula(workflow, formula)
-  workflow <- add_model(workflow, model)
+  empty_ellipses(...)
 
-  last_fit_workflow(workflow, split, metrics)
+  wflow <- add_model(workflow(), object)
+
+  if (is_recipe(preprocessor)) {
+    wflow <- add_recipe(wflow, preprocessor)
+  } else if (rlang::is_formula(preprocessor)) {
+    wflow <- add_formula(wflow, preprocessor)
+  }
+
+  last_fit_workflow(wflow, split, metrics)
 }
+
 
 #' @rdname last_fit
 #' @export