+++ R/tune_Bayes.R
@@ -132,14 +132,15 @@ tune_bayes <- function(object, ...) {
 }
 
 #' @export
-#' @rdname tune_bayes
 tune_bayes.default <- function(object, ...) {
-  stop("The first argument should be either a formula, recipe, or workflow.",
-       call. = FALSE)
+  msg <- paste0(
+    "The first argument to [tune_bayes()] should be either ",
+    "a model or workflow."
+  )
+  rlang::abort(msg)
 }
 
 #' @export
-#' @rdname tune_bayes
 tune_bayes.recipe <- function(object,
                               model,
                               resamples,
@@ -151,24 +152,18 @@ tune_bayes.recipe <- function(object,
                               initial = 5,
                               control = control_bayes()) {
 
+  lifecycle::deprecate_soft("0.1.0",
+                            what = "tune_bayes.recipe()",
+                            details = deprecate_msg(match.call(), "tune_bayes"))
   empty_ellipses(...)
 
-  if (is_missing(model) || !inherits(model, "model_spec")) {
-    stop("`model` should be a parsnip model specification object.", call. = FALSE)
-  }
-
-  wflow <-
-    workflow() %>%
-    add_recipe(object) %>%
-    add_model(model)
-
-  tune_bayes_workflow(wflow, resamples = resamples, iter = iter, param_info = param_info,
-                      metrics = metrics, objective = objective, initial = initial,
-                      control = control, ...)
+  tune_bayes(model, preprocessor = object, resamples = resamples,
+             iter = iter, param_info = param_info,
+             metrics = metrics, objective = objective,
+             initial = initial, control = control)
 }
 
 #' @export
-#' @rdname tune_bayes
 tune_bayes.formula <- function(formula,
                                model,
                                resamples,
@@ -180,21 +175,50 @@ tune_bayes.formula <- function(formula,
                                initial = 5,
                                control = control_bayes()) {
 
+  lifecycle::deprecate_soft("0.1.0",
+                            what = "tune_bayes.formula()",
+                            details = deprecate_msg(match.call(), "tune_bayes"))
   empty_ellipses(...)
 
-  if (is_missing(model) || !inherits(model, "model_spec")) {
-    stop("`model` should be a parsnip model specification object.", call. = FALSE)
+  tune_bayes(model, preprocessor = formula, resamples = resamples,
+             iter = iter, param_info = param_info,
+             metrics = metrics, objective = objective,
+             initial = initial, control = control)
+}
+
+#' @export
+#' @rdname tune_bayes
+tune_bayes.model_spec <- function(object,
+                                  preprocessor,
+                                  resamples,
+                                  ...,
+                                  iter = 10,
+                                  param_info = NULL,
+                                  metrics = NULL,
+                                  objective = exp_improve(),
+                                  initial = 5,
+                                  control = control_bayes()) {
+
+  if (rlang::is_missing(preprocessor) || !is_preprocessor(preprocessor)) {
+    rlang::abort(paste("To tune a model spec, you must preprocess",
+                       "with a formula or recipe"))
   }
 
-  wflow <-
-    workflow() %>%
-    add_formula(formula) %>%
-    add_model(model)
+  empty_ellipses(...)
+
+  wflow <- add_model(workflow(), object)
+
+  if (is_recipe(preprocessor)) {
+    wflow <- add_recipe(wflow, preprocessor)
+  } else if (rlang::is_formula(preprocessor)) {
+    wflow <- add_formula(wflow, preprocessor)
+  }
 
   tune_bayes_workflow(wflow, resamples = resamples, iter = iter, param_info = param_info,
                       metrics = metrics, objective = objective, initial = initial,
                       control = control, ...)
 }
+
 
 #' @export
 #' @rdname tune_bayes
@@ -211,10 +235,10 @@ tune_bayes.workflow <-
 
     empty_ellipses(...)
 
-  tune_bayes_workflow(object, resamples = resamples, iter = iter, param_info = param_info,
-                      metrics = metrics, objective = objective, initial = initial,
-                      control = control, ...)
-}
+    tune_bayes_workflow(object, resamples = resamples, iter = iter, param_info = param_info,
+                        metrics = metrics, objective = objective, initial = initial,
+                        control = control, ...)
+  }
 
 tune_bayes_workflow <-
   function(object, resamples, iter = 10, param_info = NULL, metrics = NULL,
@@ -223,7 +247,7 @@ tune_bayes_workflow <-
     start_time <- proc.time()[3]
 
     check_rset(resamples)
-     metrics <- check_metrics(metrics, object)
+    metrics <- check_metrics(metrics, object)
     metrics_data <- metrics_info(metrics)
     metrics_name <- metrics_data$.metric[1]
     maximize <- metrics_data$direction[metrics_data$.metric == metrics_name] == "maximize"
@@ -555,7 +579,7 @@ is_cataclysmic <- function(x) {
   is_err <- purrr::map_lgl(x$.metrics, inherits, c("simpleError", "error"))
   if (any(!is_err)) {
     is_good <- purrr::map_lgl(x$.metrics[!is_err],
-                               ~ tibble::is_tibble(.x) && nrow(.x) > 0)
+                              ~ tibble::is_tibble(.x) && nrow(.x) > 0)
     is_err[!is_err] <- !is_good
   }
   all(is_err)